[API CHANGE] lok: Don't use 'bool' and 'uint64_t' in the stable API.
This is a f70e0ec6b3c61a7c7caa469949b0ac8016c89854 follow-up.

+++ include/LibreOfficeKit/LibreOfficeKit.hxx
@@ -554,7 +554,7 @@ public:
      * @since LibreOffice 6.0
      * @see LibreOfficeKitOptionalFeatures
      */
-    void setOptionalFeatures(uint64_t features)
+    void setOptionalFeatures(unsigned long long features)
     {
         return mpThis->pClass->setOptionalFeatures(mpThis, features);
     }
